# api/openapi.yaml
openapi: 3.1.0
info:
  title: AURA — Ashesi University Resource Allocation API
  version: "1.0.0"
  description: >
    Contract-first REST API for AURA, Ashesi University's resource allocation and
    reservation platform (§8). JSON over HTTPS,
    versioned under /api/v1. Errors use application/problem+json (RFC 9457) with a
    stable machine `code`. Cursor pagination via ?limit & ?cursor. This document is
    the source of truth from which the web and mobile typed clients are generated.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.aura.ashesi.edu
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/v1/buildings:
    get:
      tags: [Catalogue]
      summary: List buildings
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Building" } }
    post:
      tags: [Catalogue]
      summary: Create a building (ROOM_MANAGE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code: { type: string }
                name: { type: string }
                campus: { type: string, nullable: true }
      responses: { "201": { description: Created }, "403": { $ref: "#/components/responses/Problem" } }
  /api/v1/buildings/{id}:
    get:
      tags: [Catalogue]
      summary: Get building detail
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Building" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/buildings/{id}/images:
    post:
      tags: [Catalogue]
      summary: Upload a building main image and gallery images (ROOM_MANAGE)
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                main: { type: string, format: binary }
                gallery:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: Updated building
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Building" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/equipment:
    get:
      tags: [Catalogue]
      summary: List equipment
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Equipment" } }
    post:
      tags: [Catalogue]
      summary: Create equipment (ROOM_MANAGE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code: { type: string }
                name: { type: string }
      responses: { "201": { description: Created }, "403": { $ref: "#/components/responses/Problem" } }
  /api/v1/equipment/{id}:
    get:
      tags: [Catalogue]
      summary: Get equipment detail
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Equipment" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/equipment/{id}/images:
    post:
      tags: [Catalogue]
      summary: Upload an equipment main image and gallery images (ROOM_MANAGE)
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                main: { type: string, format: binary }
                gallery:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Equipment" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/rooms:
    get:
      tags: [Catalogue]
      summary: List/filter rooms (FR6 filters)
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
        - { name: building_id, in: query, schema: { type: string, format: uuid } }
        - { name: min_capacity, in: query, schema: { type: integer } }
        - { name: room_type, in: query, schema: { $ref: "#/components/schemas/RoomType" } }
        - { name: equipment, in: query, schema: { type: string }, description: Comma-separated equipment codes }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Room" } }
                  next_cursor: { type: string, nullable: true }
  /api/v1/rooms/{id}:
    get:
      tags: [Catalogue]
      summary: Get room detail
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  room: { $ref: "#/components/schemas/Room" }
                  equipment:
                    type: array
                    nullable: true
                    items: { $ref: "#/components/schemas/RoomEquipment" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/rooms/{id}/images:
    post:
      tags: [Catalogue]
      summary: Upload a room main image and gallery images (ROOM_MANAGE)
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                main: { type: string, format: binary }
                gallery:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    cookieAuth: { type: apiKey, in: cookie, name: __Host-access }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Limit: { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
  responses:
    Problem:
      description: RFC 9457 problem document
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: Stable machine-readable error code }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
...
    RoomType: { type: string, enum: [LECTURE_HALL, LAB, SEMINAR_ROOM, AUDITORIUM, CONFERENCE_ROOM] }
...
    Building:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        campus: { type: string, nullable: true }
        image_url: { type: string, nullable: true }
        image_public_id: { type: string, nullable: true }
        gallery_urls: { type: array, items: { type: string } }
        gallery_public_ids: { type: array, items: { type: string } }
    Equipment:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        image_url: { type: string, nullable: true }
        image_public_id: { type: string, nullable: true }
        gallery_urls: { type: array, items: { type: string } }
        gallery_public_ids: { type: array, items: { type: string } }
    RoomEquipment:
      type: object
      properties:
        equipment_id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        image_url: { type: string, nullable: true }
        quantity: { type: integer }
    Room:
      type: object
      properties:
        id: { type: string, format: uuid }
        room_code: { type: string }
        name: { type: string }
        building_id: { type: string, format: uuid }
        capacity: { type: integer }
        room_type: { $ref: "#/components/schemas/RoomType" }
        status: { type: string }
        building_code: { type: string }
        building_name: { type: string }
        image_url: { type: string, nullable: true }
        image_public_id: { type: string, nullable: true }
        gallery_urls: { type: array, items: { type: string } }
        gallery_public_ids: { type: array, items: { type: string } }
...